Only write a deps.jl file if one doesnt' already exist

+++ deps/build.jl
@@ -27,8 +27,10 @@ if Libdl.dlopen_e("libmbedtls")    != C_NULL &&
    validate_mbed("", Libdl.dlopen_e("libmbedcrypto")) &&
    get(ENV, "FORCE_BUILD", "") != "true"
    println("Using system libraries...")
-   open("deps.jl", "w") do f
-       write(f, systemlibs)
+   if !isfile("deps.jl") && readstring("deps.jl") != systemlibs
+       open("deps.jl", "w") do f
+           write(f, systemlibs)
+       end
    end
    exit()
 end
